backend-cost/requirements.txt
boto3
flask
flask-cors
requests
python-dotenv
redis
